fix rendering images on quick change
When browsing more quickly through artists or albums the images of all
visited items were eventually added to the last selected buffer. Images
are loaded asyncronuously and so once arrived just inserted in the then
current buffer. Now the url is first inserted and the image is only
rendered if the url has not been changed since then.